Map: Fixed memory allocation related issues in insert and move.
There's 2 big issues here:

1. When inserting a key, we didn't check that our atom was still
valid. This is an issue if our insert was interupted and later
resumed. The fix is simple: check the dealloc var and realloc if
needed.

2. Memory would leak if moving an half-inserted bucket because it
would get tombstoned without being deallocated. This is trickier
because the move can't detect this nor should it. The work around is
to move half-inserted buckets as is and let the insert resume from
where it left off.

Still a segfault I need to track down though.