Tls: Added a TLS class that handles construction and destruction.
Still needs to be tested to make sure it's actually faster then
pthread_getspecific and pthread_setspecific. If nothing else it's less
quirky then those functions so I don't think this is an actual issue.

It does make use the of the pthread tls to ensure that the destructor
is invoked when the thread dies. I don't know of any other facilities
in posix or linux that allows the detection of dying thread so I
believe this can be considered a hack.