Rcu: Defer list is now implemented using List<T>
Turns out that the complexity surounding the deferList was not
required because there is no race between an exit and a defer op. The
reason is that when pushing a new node, as long as we don't
dereference head we can't read a node that was deleted by an exit op.
If the old head was deleted then we won't be able to set our node due
to the head having changed to nullptr hence another loop will be
required.