Alloc: Free is now wait-free by using a hierarchical recycled bitfield.
The main idea here is that we use that bitfield to synchronize both
kill and free on the same word using an atomic or instruction to
detect whether we're the last one to touch the page.

This does add an atomic op to the free op that is in most cases
superfluous (result is thrown away in most cases) but removes the cas
loop which is a net win.

Woo!