Alloc: Search loop starts from last successfull chunk.
This helps the scalability of the allocator when using small blocks by
preventing the search from pointlessly starting at the beginning of
the bitfield every time.

On the recycled bitfield, this ensures that we don't obsess over a
single chunk and instead gradually rotate around the bitfield. This
gives a chance for chunks to accumulate before they are transfered.